[ANCHOR-1294]: Review
